+++ api/src/main/java/com/refurbmarket/controller/OrderController.java
@@ -1,32 +1,32 @@
 package com.refurbmarket.controller;
 
-import java.util.List;
-
-import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.refurbmarket.dto.response.CategoryResponseDto;
-import com.refurbmarket.service.CategoryService;
+import com.refurbmarket.dto.request.OrderRequestDto;
+import com.refurbmarket.dto.response.OrderResponseDto;
+import com.refurbmarket.service.OrderService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 
-@Tag(name = "Category", description = "Category API")
+@Tag(name = "Order", description = "Order API")
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/categories")
-public class CategoryController {
-	private final CategoryService categoryService;
+@RequestMapping("/orders")
+public class OrderController {
+	private final OrderService orderService;
 
-	@Operation(summary = "모든 카테고리 정보 조회", description = "마켓의 모든 카테고리 정보를 조회한다.")
+	@Operation(summary = "주문 및 결제", description = "주문 후 결제를 진행한다.")
 	@ApiResponse(
-		responseCode = "200", description = "모든 카테고리 정보 조회 성공"
+		responseCode = "200", description = "주문 및 결제 성공"
 	)
-	@GetMapping
-	public List<CategoryResponseDto> getCategories() {
-		return categoryService.getCategories();
+	@PostMapping
+	public OrderResponseDto order(@RequestBody OrderRequestDto request) {
+		return orderService.order(request);
 	}
 }
