+++ api/src/main/java/com/refurbmarket/dto/PaymentType.java
@@ -1,2 +1,5 @@
-package com.refurbmarket.dto;public enum PaymentType {
+package com.refurbmarket.dto;
+
+public enum PaymentType {
+	CARD, PAY, BANK_TRANSFER
 }
