+++ api/src/main/java/com/refurbmarket/dto/request/OrderRequestDto.java
@@ -1,2 +1,18 @@
-package com.refurbmarket.dto.request;public class OrderRequestDto {
+package com.refurbmarket.dto.request;
+
+import java.util.List;
+
+import com.refurbmarket.dto.PaymentType;
+
+import lombok.Getter;
+
+@Getter
+public class OrderRequestDto {
+	private Long userId;
+	private String receiverName;
+	private String receiverAddress;
+	private String receiverPhone;
+	private List<Long> couponIssueIdList;
+	private List<OrderItemRequestDto> orderItemList;
+	private PaymentType type;
 }
