+++ api/src/main/java/com/refurbmarket/service/OrderService.java
@@ -1,48 +1,25 @@
 package com.refurbmarket.service;
 
-import static java.util.stream.Collectors.toList;
-import static java.util.stream.Collectors.toMap;
-
-import java.util.Comparator;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-
 import org.springframework.stereotype.Service;
 
-import com.refurbmarket.domain.Category;
-import com.refurbmarket.dto.response.CategoryResponseDto;
-import com.refurbmarket.repository.MyBatisCategoryRepository;
+import com.refurbmarket.dto.request.OrderRequestDto;
+import com.refurbmarket.dto.response.OrderResponseDto;
 
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-public class CategoryService {
-	private final MyBatisCategoryRepository categoryRepository;
+public class OrderService {
+	public OrderResponseDto order(OrderRequestDto dto) {
+		// 쿠폰이 유효한지 확인
 
-	public List<CategoryResponseDto> getCategories() {
-		List<Category> categories = categoryRepository.findAll();
-		Map<Long, CategoryResponseDto> dtoMap = createDtoMap(categories);
-		return createTreeStructure(categories, dtoMap);
-	}
+		// 주문할 상품 리스트 조회
 
-	private List<CategoryResponseDto> createTreeStructure(List<Category> categories,
-		Map<Long, CategoryResponseDto> dtoMap) {
-		return categories.stream()
-			.peek(category -> {
-				CategoryResponseDto dto = dtoMap.get(category.getId());
-				Optional.ofNullable(category.getParentId())
-					.ifPresent(parentId -> dtoMap.get(parentId).getChildren().add(dto));
-			})
-			.filter(category -> category.getParentId() == null)
-			.sorted(Comparator.comparing(Category::getName))
-			.map(category -> dtoMap.get(category.getId()))
-			.collect(toList());
-	}
+		// 주문할 상품 재고 확인
 
-	private Map<Long, CategoryResponseDto> createDtoMap(List<Category> categories) {
-		return categories.stream()
-			.collect(toMap(Category::getId, CategoryResponseDto::of));
+		// 주문 생성
+
+		// 결제
+		return null;
 	}
 }
